# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gapak API
  version: 1.0.0
  description: |
    Gapak is a privacy-first social network built around trust, layered identity,
    controlled visibility, secure media, live experiences, and time-based social interaction.

    Privacy is a core mechanic rather than a decorative setting. The API is designed for
    trust-aware access rules, metadata minimization, secure media delivery, and live social primitives.
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessCookie:
      type: apiKey
      in: cookie
      name: gapak_at
  schemas:
...
    RegisterRequest:
      type: object
      required: [username, displayName, password]
      properties:
        email: { type: string, format: email }
        username: { type: string }
        displayName: { type: string }
        password: { type: string }
        preferAnonymous: { type: boolean }
        deviceName: { type: string }
        deviceFingerprint: { type: string }
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login: { type: string }
        password: { type: string }
        totpCode: { type: string }
        deviceName: { type: string }
        deviceFingerprint: { type: string }
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register account with pseudonymous-friendly flow
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201': { description: Auth session created }
  /api/v1/auth/register-anonymous:
    post:
      tags: [Auth]
      summary: Register anonymity-first account without public email identity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201': { description: Anonymous auth session created }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login account without exposing private identity to client-visible flows
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: Auth session created }
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token
      parameters:
        - in: header
          name: X-CSRF-Token
          schema: { type: string }
      responses:
        '200': { description: New access token plus rotated HttpOnly refresh cookie; refresh token is never present in the JSON body }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }, { accessCookie: [] }]
      summary: Logout current or all sessions
      responses:
        '200': { description: Logout completed }
  /api/v1/auth/oauth/{provider}:
    get:
      tags: [Auth]
      summary: Return the provider authorization URL
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string, enum: [google, github, facebook] }
      responses:
        '200': { description: OAuth authorization URL }
  /api/v1/auth/callback/{provider}:
    get:
      tags: [Auth]
      summary: Complete the OAuth callback, establish the server session, then redirect to the configured frontend origin
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string, enum: [google, github, facebook] }
        - in: query
          name: code
          required: true
          schema: { type: string }
        - in: query
          name: state
          required: true
          schema: { type: string }
      responses:
        '307': { description: Redirect to the configured frontend origin after session establishment }
        '400': { description: Invalid OAuth state or missing authorization code }
...
  /api/v1/auth/2fa/setup:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      summary: Create TOTP setup secret
      responses:
        '200': { description: Secret and OTP Auth URL }
  /api/v1/auth/2fa/verify:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      summary: Verify and enable TOTP
      responses:
        '200': { description: 2FA enabled }
...